Autorização implementada, Policies

+++ app/Http/Controllers/EspecialidadeController.php
@@ -7,43 +7,57 @@ use Illuminate\Http\Request;
 
 class EspecialidadeController extends Controller
 {
-
     public function index(Request $request)
     {
-        // dd($request->all());
+        $this->authorize('viewAny', Especialidade::class);
+
         $query = Especialidade::withCount('medicos')->latest();
+
         if($request->filled('nome')){
-            $query->where('nome', 'like', "%" . $request->nome ."%");
+            $query->where('nome', 'like', "%" . $request->nome . "%");
         }
+
         $especialidades = $query->paginate(10)->appends($request->all());
+
         return inertia('Especialidade', ['especialidades' => $especialidades]);
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Especialidade::class);
+
         $validated = $request->validate([
             'nome' => 'required|string|max:50|unique:especialidades,nome'
         ]); 
 
         Especialidade::create($validated);
+
         return redirect()->route('especialidades.index');
     }
 
     public function update(Request $request, Especialidade $especialidade)
     {
-         $validated = $request->validate([
+        $this->authorize('update', $especialidade);
+
+        $validated = $request->validate([
             'nome' => 'required|string|max:50|unique:especialidades,nome,' . $especialidade->id
         ]);
+
         $especialidade->update($validated);
+
         return redirect()->route('especialidades.index');
     }
 
     public function destroy(Especialidade $especialidade)
     {
+        $this->authorize('delete', $especialidade);
+
         if($especialidade->medicos()->count() > 0){
             return redirect()->back();
         }
+
         $especialidade->delete();
+
         return redirect()->route('especialidades.index');
     }
 }
